Move react logic into its own package.
