+++ packages/core/src/Aesthetic.tsx
@@ -1,31 +1,18 @@
 /* eslint-disable max-classes-per-file, react/no-multi-comp */
 
-import React, { useState, useLayoutEffect } from 'react';
-import uuid from 'uuid/v4';
-import hoistNonReactStatics from 'hoist-non-react-statics';
 import deepMerge from 'extend';
-import { Omit } from 'utility-types';
 import isObject from './helpers/isObject';
 import stripClassPrefix from './helpers/stripClassPrefix';
 import StyleSheetManager from './StyleSheetManager';
 import UnifiedSyntax from './UnifiedSyntax';
 import {
   ClassName,
-  ClassNameGenerator,
   GlobalSheetDefinition,
   SheetMap,
-  StyledComponentClass,
   StyleName,
   StyleSheet,
   StyleSheetDefinition,
   ThemeName,
-  WithStylesOptions,
-  WithStylesProps,
-  WithStylesState,
-  WithStylesWrapperProps,
-  WithThemeOptions,
-  WithThemeProps,
-  WithThemeWrapperProps,
 } from './types';
 
 export interface AestheticOptions {
@@ -164,6 +151,22 @@ export default abstract class Aesthetic<
   flushStyles(styleName: StyleName) {}
 
   /**
+   * Return a theme object or throw an error.
+   */
+  getTheme(name?: ThemeName): Theme {
+    const themeName = name || this.options.theme;
+    const theme = this.themes[themeName];
+
+    if (__DEV__) {
+      if (!theme || !isObject(theme)) {
+        throw new Error(`Theme "${themeName}" does not exist.`);
+      }
+    }
+
+    return theme;
+  }
+
+  /**
    * Register a theme with a pre-defined set of theme settings.
    */
   registerTheme<T>(
@@ -181,6 +184,31 @@ export default abstract class Aesthetic<
 
     this.themes[themeName] = theme;
     this.globals[themeName] = this.validateDefinition(themeName, globalSheet, this.globals);
+
+    return this;
+  }
+
+  /**
+   * Set a style sheet definition for a component.
+   */
+  setStyleSheet(
+    styleName: StyleName,
+    styleSheet: StyleSheetDefinition<Theme, any>,
+    extendFrom: StyleName = '',
+  ): this {
+    if (extendFrom) {
+      if (__DEV__) {
+        if (!this.styles[extendFrom]) {
+          throw new Error(`Cannot extend from "${extendFrom}" as those styles do not exist.`);
+        } else if (extendFrom === styleName) {
+          throw new Error('Cannot extend styles from itself.');
+        }
+      }
+
+      this.parents[styleName] = extendFrom;
+    }
+
+    this.styles[styleName] = this.validateDefinition(styleName, styleSheet, this.styles);
 
     return this;
   }
@@ -220,154 +248,6 @@ export default abstract class Aesthetic<
   };
 
   /**
-   * Hook within a component to provide a style sheet.
-   */
-  useStyles<T>(
-    styleSheet: StyleSheetDefinition<Theme, T>,
-    customName: string = 'Component',
-  ): [SheetMap<ParsedBlock>, ClassNameGenerator<NativeBlock, ParsedBlock>, string] {
-    const [styleName] = useState(() => {
-      const name = `${customName}-${uuid()}`;
-
-      this.setStyleSheet(name, styleSheet);
-
-      return name;
-    });
-
-    const sheet = this.createStyleSheet(styleName);
-
-    // Flush styles on mount
-    useLayoutEffect(() => {
-      this.flushStyles(styleName);
-    }, [styleName]);
-
-    return [sheet, this.transformStyles, styleName];
-  }
-
-  /**
-   * Hook within a component to provide the current theme object.
-   */
-  useTheme(): Theme {
-    return this.getTheme();
-  }
-
-  /**
-   * Wrap a React component with an HOC that injects the defined style sheet as a prop.
-   */
-  withStyles<T>(
-    styleSheet: StyleSheetDefinition<Theme, T>,
-    options: WithStylesOptions = {},
-  ) /* infer */ {
-    const aesthetic = this;
-    const {
-      extendable = this.options.extendable,
-      extendFrom = '',
-      passThemeProp = this.options.passThemeProp,
-      pure = this.options.pure,
-      stylesPropName = this.options.stylesPropName,
-      themePropName = this.options.themePropName,
-    } = options;
-
-    return function withStylesFactory<Props extends object = {}>(
-      WrappedComponent: React.ComponentType<Props & WithStylesProps<Theme, ParsedBlock>>,
-    ): StyledComponentClass<Theme, Props & WithStylesWrapperProps> {
-      const baseName = WrappedComponent.displayName || WrappedComponent.name;
-      const styleName = `${baseName}-${uuid()}`;
-      const Component = pure ? React.PureComponent : React.Component;
-
-      type OwnState = WithStylesState<Props, ParsedBlock>;
-
-      aesthetic.setStyleSheet(styleName, styleSheet, extendFrom);
-
-      class WithStyles extends Component<Props & WithStylesWrapperProps, OwnState> {
-        static displayName = `withStyles(${baseName})`;
-
-        static styleName = styleName;
-
-        static WrappedComponent = WrappedComponent;
-
-        static extendStyles<ET>(
-          customStyleSheet: StyleSheetDefinition<Theme, ET>,
-          extendOptions: Omit<WithStylesOptions, 'extendFrom'> = {},
-        ) {
-          if (__DEV__) {
-            if (!extendable) {
-              throw new Error(`${baseName} is not extendable.`);
-            }
-          }
-
-          return aesthetic.withStyles(customStyleSheet, {
-            ...options,
-            ...extendOptions,
-            extendFrom: styleName,
-          })(WrappedComponent);
-        }
-
-        state: OwnState = {
-          styles: aesthetic.createStyleSheet(styleName),
-        };
-
-        componentDidMount() {
-          aesthetic.flushStyles(styleName);
-        }
-
-        render() {
-          const { wrappedRef, ...props } = this.props;
-          const extraProps: WithStylesProps<Theme, ParsedBlock> = {
-            [stylesPropName as 'styles']: this.state.styles,
-            ref: wrappedRef,
-          };
-
-          if (passThemeProp) {
-            extraProps[themePropName as 'theme'] = aesthetic.getTheme();
-          }
-
-          return <WrappedComponent {...props as any} {...extraProps} />;
-        }
-      }
-
-      hoistNonReactStatics(WithStyles, WrappedComponent);
-
-      return WithStyles;
-    };
-  }
-
-  /**
-   * Wrap a React component with an HOC that injects the current theme object as a prop.
-   */
-  withTheme(options: WithThemeOptions = {}) /* infer */ {
-    const aesthetic = this;
-    const { pure = this.options.pure, themePropName = this.options.themePropName } = options;
-
-    return function withThemeFactory<Props extends object = {}>(
-      WrappedComponent: React.ComponentType<Props & WithThemeProps<Theme>>,
-    ): React.ComponentClass<Props & WithThemeWrapperProps> {
-      const baseName = WrappedComponent.displayName || WrappedComponent.name;
-      const Component = pure ? React.PureComponent : React.Component;
-
-      class WithTheme extends Component<Props & WithThemeWrapperProps> {
-        static displayName = `withTheme(${baseName})`;
-
-        static WrappedComponent = WrappedComponent;
-
-        render() {
-          const { wrappedRef, ...props } = this.props;
-          const extraProps: WithThemeProps<Theme> = {
-            [themePropName as 'theme']: aesthetic.getTheme(),
-            ref: wrappedRef,
-          };
-
-          return <WrappedComponent {...props as any} {...extraProps} />;
-        }
-      }
-
-      hoistNonReactStatics(WithTheme, WrappedComponent);
-
-      return WithTheme;
-    };
-  }
-
-  /**
    * Retrieve the defined component styles. If the definition is a function,
    * execute it while passing the current theme and React props.
    */
@@ -398,22 +278,6 @@ export default abstract class Aesthetic<
   }
 
   /**
-   * Return a theme object or throw an error.
-   */
-  protected getTheme(customTheme: ThemeName = ''): Theme {
-    const themeName = customTheme || this.options.theme;
-    const theme = this.themes[themeName];
-
-    if (__DEV__) {
-      if (!theme || !isObject(theme)) {
-        throw new Error(`Theme "${themeName}" does not exist.`);
-      }
-    }
-
-    return theme;
-  }
-
-  /**
    * Process from an Aesthetic style sheet to an adapter native style sheet.
    */
   protected processStyleSheet(
@@ -422,31 +286,6 @@ export default abstract class Aesthetic<
   ): SheetMap<ParsedBlock> {
     // @ts-ignore Allow spread
     return { ...styleSheet };
-  }
-
-  /**
-   * Set a style sheet definition for a component.
-   */
-  protected setStyleSheet(
-    styleName: StyleName,
-    styleSheet: StyleSheetDefinition<Theme, any>,
-    extendFrom: StyleName = '',
-  ): this {
-    if (extendFrom) {
-      if (__DEV__) {
-        if (!this.styles[extendFrom]) {
-          throw new Error(`Cannot extend from "${extendFrom}" as those styles do not exist.`);
-        } else if (extendFrom === styleName) {
-          throw new Error('Cannot extend styles from itself.');
-        }
-      }
-
-      this.parents[styleName] = extendFrom;
-    }
-
-    this.styles[styleName] = this.validateDefinition(styleName, styleSheet, this.styles);
-
-    return this;
   }
 
   /**
