+++ packages/react/README.md
@@ -1,0 +1,19 @@
+# Aesthetic via React
+
+[![Build Status](https://travis-ci.org/milesj/aesthetic.svg?branch=master)](https://travis-ci.org/milesj/aesthetic)
+[![npm version](https://badge.fury.io/js/aesthetic-react.svg)](https://www.npmjs.com/package/aesthetic-react)
+[![npm deps](https://david-dm.org/milesj/aesthetic.svg?path=packages/react)](https://www.npmjs.com/package/aesthetic-react)
+
+TODO
+
+## Requirements
+
+- React 16.3+
+
+## Installation
+
+```
+yarn add aesthetic aesthetic-react
+// Or
+npm install aesthetic aesthetic-react
+```
