+++ packages/react/package.json
@@ -1,0 +1,37 @@
+{
+  "name": "aesthetic-react",
+  "version": "1.0.0",
+  "description": "React hooks, composers, and more for Aesthetic.",
+  "keywords": [
+    "aesthetic",
+    "react",
+    "styles"
+  ],
+  "repository": "https://github.com/milesj/aesthetic/tree/master/packages/react",
+  "license": "MIT",
+  "main": "./lib/index.js",
+  "module": "./esm/index.js",
+  "types": "./lib/index.d.ts",
+  "sideEffects": false,
+  "publishConfig": {
+    "access": "public"
+  },
+  "dependencies": {
+    "@types/hoist-non-react-statics": "^3.3.1",
+    "@types/prop-types": "^15.7.1",
+    "@types/react": "^16.8.18",
+    "@types/uuid": "^3.4.4",
+    "hoist-non-react-statics": "^3.3.0",
+    "prop-types": "^15.7.2",
+    "utility-types": "^3.7.0",
+    "uuid": "^3.3.2"
+  },
+  "peerDependencies": {
+    "aesthetic": "^3.0.0",
+    "react": "^16.3.0"
+  },
+  "devDependencies": {
+    "react": "^16.8.6",
+    "react-dom": "^16.8.6"
+  }
+}
