+++ packages/react/src/index.ts
@@ -1,0 +1,11 @@
+/**
+ * @copyright   2017-2019, Miles Johnson
+ * @license     https://opensource.org/licenses/MIT
+ */
+
+import useStylesFactory from './useStylesFactory';
+import useThemeFactory from './useThemeFactory';
+import withStylesFactory from './withStylesFactory';
+import withThemeFactory from './withThemeFactory';
+
+export { useStylesFactory, useThemeFactory, withStylesFactory, withThemeFactory };
